cc: refactor DataModifier for multiple-backend framework
